Make traverser cost float, not integer.

+++ Scripts/AI/Pathfinding/ITraverser.cs
@@ -5,6 +5,6 @@ namespace Claw.AI.Pathfinding {
         bool CanEndOn(T node);
         bool CanTraverse(T node);
         bool AddToResult(T node);
-        int GetTraverseCost(T start, T end);
+        float GetTraverseCost(T start, T end);
     }
 }
